# storage/api-docs/api-docs-admin-v1.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation for admin",
        "description": "L5 Swagger OpenApi description for admin",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/admin/v1/get-products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "operationId": "getAllProducts",
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "object"
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "APIKey": []
                    }
                ]
            }
        },
        "/admin/v1/store-product": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "new version",
                "operationId": "ed265f16f0b11d3dfa7ee98cbc95807a",
                "requestBody": {
                    "description": "Body request needed to update activity weekdays",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "title"
                                    },
                                    "language": {
                                        "description": "lang"
                                    },
                                    "product_subtitle": {
                                        "description": "subtitle",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "description",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "price",
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "image": {
                                        "description": "file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "object"
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKey": []
                    }
                ]
            }
        },
        "/admin/v1/update-product-language/{product_slug}": {
            "put": {
                "tags": [
                    "Product"
                ],
                "operationId": "1cd721145ceebde967c408cd265ffc45",
                "parameters": [
                    {
                        "name": "product_slug",
                        "in": "path",
                        "description": "the product to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "x-language",
                        "in": "header",
                        "description": "the product to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Body request needed to update activity weekdays",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_subtitle": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "object"
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "APIKey": {
                "type": "apiKey",
                "name": "X-Authorization",
                "in": "header"
            }
        }
    }
}